Split CLI methods in relevant classes. Added cache clearing directory removal securing
