+++ core/config/commands.json
@@ -2,17 +2,17 @@
   "debug": [
     {
       "name": "container",
-      "method": "Helper\\Majordomo::container",
+      "method": "Helper\\ServiceContainerCLI::container",
       "man": "List all services declared in the Container"
     },
     {
       "name": "config",
-      "method": "Helper\\Majordomo::config",
+      "method": "Helper\\ConfigurationManagerCLI::config",
       "man": "List App Configuration"
     },
     {
       "name": "router",
-      "method": "Helper\\Majordomo::router",
+      "method": "Helper\\RouterCLI::router",
       "man": "List routes and details"
     }
   ],
